requirements.txt
python-dotenv
fastapi
uvicorn
openai
langchain
chromadb
tiktoken
pypdf
transformers
sentence_transformers
bitsandbytes
InstructorEmbedding